WIP re-organizing client code
