+++ clients/jvm/client-api/src/main/kotlin/com/evidentdb/client/codec.kt
