+++ clients/jvm/client-api/src/main/kotlin/io/cloudevents/protobuf/adapter.kt
