+++ clients/jvm/client-api/src/main/proto/cloudevents.proto
@@ -1,0 +1,1 @@
+../../../../../../interface/cloudevents.proto
