+++ clients/jvm/client-api/src/main/proto/domain.proto
@@ -1,0 +1,1 @@
+../../../../../../interface/domain.proto
