+++ clients/jvm/client-api/src/main/proto/service.proto
@@ -1,0 +1,1 @@
+../../../../../../interface/service.proto
