+++ clients/jvm/client-java-caching/src/test/kotlin/com/evidentdb/client/kotlin/test/QueryTests.kt
