--- clients/jvm/src/main/proto/cloudevents.proto
@@ -1,1 +1,0 @@
-../../../../../interface/cloudevents.proto
