--- clients/jvm/src/main/proto/domain.proto
@@ -1,1 +1,0 @@
-../../../../../interface/domain.proto
