--- clients/jvm/src/main/proto/service.proto
@@ -1,1 +1,0 @@
-../../../../../interface/service.proto
